+++ nautilus_trader/data/client.pyx
@@ -156,7 +156,7 @@ cdef class DataClient(Component):
     def _handle_data_py(self, Data data):
         self._handle_data(data)
 
-    def _handle_data_response_py(self, DataType data_type, Data data, UUID4 correlation_id):
+    def _handle_data_response_py(self, DataType data_type, object data, UUID4 correlation_id):
         self._handle_data_response(data_type, data, correlation_id)
 
 # -- DATA HANDLERS ---------------------------------------------------------------------------------
@@ -164,7 +164,7 @@ cdef class DataClient(Component):
     cpdef void _handle_data(self, Data data) except *:
         self._msgbus.send(endpoint="DataEngine.process", msg=data)
 
-    cpdef void _handle_data_response(self, DataType data_type, Data data, UUID4 correlation_id) except *:
+    cpdef void _handle_data_response(self, DataType data_type, object data, UUID4 correlation_id) except *:
         cdef DataResponse response = DataResponse(
             client_id=self.id,
             data_type=data_type,